requirements.txt
click==8.0.4
requests==2.27.1
google-auth==2.6.2